+++ MalieDatPacker/malie_encryption.h
@@ -11,8 +11,8 @@ public:
 	MalieEncryption(const MalieKey& key) :m_key(key) {
 
 	}
-	void encrypt(unsigned char* data, unsigned int data_length, unsigned int offset) const override;
-	void decrypt(unsigned char* data, unsigned int data_length, unsigned int offset) const override;
+	void encrypt(unsigned char* data, size_t data_length, size_t offset) const override;
+	void decrypt(unsigned char* data, size_t data_length, size_t offset) const override;
 private:
 	const MalieKey m_key;
 };
