+++ MalieDatPacker/utils.h
@@ -11,7 +11,7 @@ void ToShiftJis(const std::wstring& str, char* dst, size_t dst_size);
 
 void cut_file(const wchar_t* filename, const wchar_t* output_filename, size_t file_size);
 
-int file_compare(const wchar_t* filepath1, const wchar_t* filepath2);
+long long file_compare(const wchar_t* filepath1, const wchar_t* filepath2);
 
 #ifndef UTILS_H
 #define UTILS_H
@@ -25,7 +25,7 @@ template<typename T>
 int findPosition(const std::vector<T>& vec, const T& target) {
 	auto it = std::find(vec.begin(), vec.end(), target);
 	if (it != vec.end()) {
-		return std::distance(vec.begin(), it);
+		return static_cast<int>(std::distance(vec.begin(), it));
 	}
 	return -1;
 }
